feat: add StartingPrice component
